requirements.txt
# requirements for DIY wireless Temp/Humidity sensor solution
spidev
pycrypto
django<1.9
django-tastypie
uwsgi
whisper
cairocffi<0.7
django-tagging
pyparse
pytz
ipaddress
